메뉴등록 백엔드 완료_메뉴포스까지

+++ src/main/java/com/Iot7_1team/pos_web/dto/MenuRegisterRequestDTO.java
@@ -9,8 +9,8 @@ public class MenuRegisterRequestDTO {
     private String posLoginId;
     private String menuName;
     private String category;
-    private int price;
-    private int calorie;
+    private Double price;
+    private Double calorie;
     private String ingredients;
     private boolean dietYn;
 }
